requirements.txt
bcrypt==5.0.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
idna==3.13
invoke==3.0.3
lxml==6.1.0
ncclient==0.7.1
paramiko==4.0.0
pycparser==3.0
PyNaCl==1.6.2
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
urllib3==2.6.3
